Created Play Mode (for playing with the Agent).
Implemented a new class GameAgent (game.py) in order to allow a player
to play a game agaisnt the trained Agent. This class needs a path to
the agents weights (.h5) and will use this agent as the oponent.

Also updated Agent (agent.py) in order to select moves in play mode i.e.
only using the neural net (not MCTS).