Bugfix and features
Changed prediction worker communication to sockets. This will make the
module more extensible and network-distributable.

Also, the code has been debugged to fix the high memory usage. This was
mainly produced by a prior update of the fully expanded nodes on all
iterations, when this has to be once executed.

Times per move with 900 iterations are about 30sec. with 6 search
threads.